Python/localizationAPI/requirements.txt
Django==3.1
djangorestframework==3.11.1
docker==4.3.1
psycopg2-binary==2.8.6
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
djangorestframework-simplejwt==4.6.0
pyjwt==2.0.1
